Implements basic tab structure
