+++ screens/GardenTabScreen.tsx
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+import { Text, View } from "../components/Themed";
+
+export const GardenTabScreen = () => {
+  return <View style={styles.container}></View>;
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+  },
+  separator: {
+    marginVertical: 30,
+    height: 1,
+    width: "80%",
+  },
+});
