--- screens/TabOneScreen.tsx
@@ -1,28 +1,0 @@
-import { StyleSheet } from "react-native";
-import { Text, View } from "../components/Themed";
-import { RootTabScreenProps } from "../types";
-import { AddBedField } from "../components/AddBedField";
-import { GardenBedsList } from "../components/GardenBedsList";
-export default function TabOneScreen({
-  navigation,
-}: RootTabScreenProps<"TabOne">) {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Garden Beds</Text>
-      <GardenBedsList />
-      <AddBedField />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-});
